+++ src/Payone/Request/Debit/AbstractDebitAuthorizeRequest.php
@@ -37,7 +37,7 @@ abstract class AbstractDebitAuthorizeRequest
             'iban'              => $iban,
             'bic'               => $bic,
             'bankaccountholder' => $accountOwner,
-            'amount'            => (int) ($transaction->getOrder()->getAmountTotal() * (10 ** $currency->getDecimalPrecision())),
+            'amount'            => (int) round(($transaction->getOrder()->getAmountTotal() * (10 ** $currency->getDecimalPrecision()))),
             'currency'          => $currency->getIsoCode(),
             'reference'         => $transaction->getOrder()->getOrderNumber(),
         ];