+++ tests/cypress/commands/finish_checkout.js
@@ -1,0 +1,5 @@
+Cypress.Commands.add('finishCheckout', () => {
+    cy.get('#tos').check({force: true});
+    cy.get('#confirmFormSubmit').click({force: true});
+    cy.url().should('include', 'checkout/finish')
+})